mjpeg: CB1 fix mjpeg report timeout frequently. [1/1]
PD#SWPL-141638

Problem:
start_process_time is a 32bit var, but it's set
ulong 64bit jiffies value. timeout is reported
without 200ms which set by decode_timeout_val.

Solution:
change start_process_time from u32 to ulong

Verify:
AN400-T7

+++ drivers/frame_provider/decoder/mjpeg/vmjpeg_multi.c
@@ -214,7 +214,7 @@ struct vdec_mjpeg_hw_s {
 	struct firmware_s *fw;
 	struct timer_list check_timer;
 	u32 decode_timeout_count;
-	u32 start_process_time;
+	ulong start_process_time;
 	u32 last_vld_level;
 	u8 eos;
 	u32 frame_num;
