+++ drivers/frame_provider/decoder_v4l/mjpeg/vmjpeg_multi.c
@@ -222,7 +222,7 @@ struct vdec_mjpeg_hw_s {
 	struct firmware_s *fw;
 	struct timer_list check_timer;
 	u32 decode_timeout_count;
-	u32 start_process_time;
+	ulong start_process_time;
 	u32 last_vld_level;
 	bool eos;
 	u32 frame_num;
